requirements (1).txt
streamlit
deepface
numpy
pillow
opencv-python-headless
tensorflow
